Add javafx list subscription implementation
