+++ build.gradle
@@ -21,5 +21,6 @@ allprojects {
     repositories {
         mavenCentral()
         jcenter()
+        maven { url 'https://jitpack.io' }
     }
 }
