+++ integration/javafx/build.gradle
@@ -2,8 +2,13 @@ apply from: rootProject.file('gradle/jvm.gradle')
 
 dependencies {
     api project(':miop-jvm')
+    api 'com.github.jcornaz.collekt:collekt-core-jvm8:0.0.2'
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-javafx:$coroutines_version"
 
     testCompile project(':miop-internal-test-jvm')
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
 }
